requirements.txt
bcrypt>=4.3.0
certifi>=2025.8.3
cffi>=1.17.1
charset-normalizer>=3.4.3
cryptography>=45.0.7
idna>=3.10
invoke>=2.2.0
markdown-it-py>=4.0.0
mdurl>=0.1.2
netmiko>=4.6.0
ntc_templates>=8.0.0
numpy>=2.3.2
paramiko>=4.0.0
pillow>=11.3.0
prompt_toolkit>=3.0.52
psutil>=7.0.0
pycparser>=2.22
pydub>=0.25.1
Pygments>=2.19.2
pymssql>=2.3.7
PyNaCl>=1.5.0
pyodbc>=5.2.0
pyserial>=3.5
pyspnego>=0.12.0
PyYAML>=6.0.2
requests>=2.32.5
requests_ntlm>=1.3.0
rich>=14.1.0
ruamel.yaml>=0.18.15
ruamel.yaml.clib>=0.2.12
scapy>=2.6.1
scp>=0.15.0
simpleaudio>=1.0.4
sounddevice>=0.5.2
textfsm>=2.1.0
tftpy>=0.8.6
urllib3>=2.5.0
wcwidth>=0.2.13
